imported print_files_with_one_replica_on_certain_pool
print_files_with_one_replica_on_certain_pool is a program which queries Chimera
for files for which exactly one replica exists in a dCache cluster and when that
is found on a specified pool.
The list of files is printed in one of several output formats.

This is especially helpful, when a pool has been lost and a list of all lost
files is needed.

It should be noted, that dCache’s storage file flags (for example “precious”,
“cached” or “sticky”) are not taken into account. This means that also
cached-only lost files will be reported, which are files that were neither
“precious” nor “cached” and “sticky”. Further, any replicas on tape are not
taken into account.

Limitations:
Currently only PostgreSQL (using psql) is supported.
Database settings cannot be specified and are neither automatically determined
from dCache’s configuration (if available). This means that Chimera must use the
database name “chimera” and access to it must be possible via the username
“chimera”.

Acknowledgements go to Xavier Mol <xavier.mol@kit.edu> for help with the SQL
queries.

• Imported the initial version of print_files_with_one_replica_on_certain_pool.

Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>