imported dcache_admin
dcache_admin is helper program to connect to dCache’s Administration Shell.

It takes commands either via its arguments or via standard intput and tries to
detect whether it is connected to a terminal or not.

A dCache service “admin” is expected to run on the same host and the port is
detected via the names “dcache-admin” respectively “dcache-admin-batch” in the
internet network services list “/etc/services”.
Passwordless SSH login via the public key authentication method must be
configured.

• Imported the initial version of dcache_admin.

Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>